+++ sdk/tables/azure-data-tables/azure/data/tables/_table_client.py
@@ -32,7 +32,7 @@ from ._models import TableEntityPropertiesPaged, UpdateMode, AccessPolicy
 
 EntityType = Union[TableEntity, Mapping[str, Any]]
 OperationType = str  # Literal['create', 'delete', 'update', 'upsert']
-BatchOperationType = Union[Tuple[EntityType, OperationType], Tuple[EntityType, OperationType, Mapping[str, Any]]]
+TransactionOperationType = Union[Tuple[EntityType, OperationType], Tuple[EntityType, OperationType, Mapping[str, Any]]]
 
 
 class TableClient(TablesBaseClient):
@@ -614,18 +614,24 @@ class TableClient(TablesBaseClient):
         except HttpResponseError as error:
             _process_table_error(error)
 
-    def send_batch(
+    def submit_transaction(
         self,
-        batch,  # type: Iterable[BatchOperationType]
+        operations,  # type: Iterable[TransactionOperationType]
         **kwargs  # type: Any
     ):
         # type: (...) -> List[Tuple[TableEntity, Mapping[str, Any]]]
-        """Commit a TableBatchOperations to send requests to the server
+        """Commit a list of operations as a single transaction.
 
+        If any one of these operations fails, the entire transaction will be rejected.
+
+        :param operations: The list of operations to commit in a transaction. This should be a list of
+         tuples containing an operation name, the entity on which to operate, and optionally, a dict of additional
+         kwargs for that operation.
+        :type operations: Iterable[Tuple[str, EntityType]]
         :return: A list of tuples, each containing the entity operated on, and a dictionary
          of metadata returned from the service.
-        :rtype: List[Tuple[Mapping[str, Any], Mapping[str, Any]]]
-        :raises ~azure.data.tables.BatchErrorException:
+        :rtype: List[Tuple[TableEntity, Mapping[str, Any]]]
+        :raises ~azure.data.tables.TableTransactionError:
 
         .. admonition:: Example:
 
@@ -634,7 +640,7 @@ class TableClient(TablesBaseClient):
                 :end-before: [END batching]
                 :language: python
                 :dedent: 8
-                :caption: Using batches to send multiple requests at once
+                :caption: Using transactions to send multiple requests at once
         """
         batched_requests = TableBatchOperations(
             self._client,
@@ -644,7 +650,7 @@ class TableClient(TablesBaseClient):
             self.table_name,
             **kwargs
         )
-        for operation in batch:
+        for operation in operations:
             try:
                 operation_kwargs = operation[2]
             except IndexError: