+++ sdk/tables/azure-data-tables/samples/sample_batching.py
@@ -10,7 +10,7 @@
 FILE: sample_batching.py
 
 DESCRIPTION:
-    These samples demonstrate how to use the batching API to perform multiple
+    These samples demonstrate how to use the batch transaction API to perform multiple
     operations within a single request
 
 USAGE:
@@ -41,9 +41,9 @@ class CreateClients(object):
             self.access_key,
             self.endpoint
         )
-        self.table_name = "sampleBatching"
+        self.table_name = "sampleTransaction"
 
-    def sample_batching(self):
+    def sample_transaction(self):
         # Instantiate a TableServiceClient using a connection string
         entity1 = {
             'PartitionKey': 'pk001',
@@ -75,7 +75,7 @@ class CreateClients(object):
         }
 
         # [START batching]
-        from azure.data.tables import TableClient, UpdateMode, BatchErrorException
+        from azure.data.tables import TableClient, UpdateMode, TableTransactionError
         from azure.core.exceptions import ResourceExistsError
         self.table_client = TableClient.from_connection_string(
             conn_str=self.connection_string, table_name=self.table_name)
@@ -86,21 +86,21 @@ class CreateClients(object):
         except ResourceExistsError:
             print("Table already exists")
 
-        self.table_client.create_entity(entity2)
-        self.table_client.create_entity(entity3)
-        self.table_client.create_entity(entity4)
+        self.table_client.upsert_entity(entity2)
+        self.table_client.upsert_entity(entity3)
+        self.table_client.upsert_entity(entity4)
 
-        batch = [
-            ('create', entity1),
+        operations = [
+            ('upsert', entity1),
             ('delete', entity2),
             ('upsert', entity3),
             ('update', entity4, {'mode': UpdateMode.REPLACE})
         ]
         try:
-            self.table_client.send_batch(batch)
-        except BatchErrorException as e:
-            print("There was an error with the batch operation")
-            print("Error: {}".format(e))
+            self.table_client.submit_transaction(operations)
+        except TableTransactionError as e:
+            print("There was an error with the transaction operation")
+            print(e)
         # [END batching]
 
     def clean_up(self):
@@ -112,6 +112,6 @@ if __name__ == '__main__':
     if sys.version_info > (3, 5):
         sample = CreateClients()
         try:
-            sample.sample_batching()
+            sample.sample_transaction()
         finally:
             sample.clean_up()