+++ sdk/tables/azure-data-tables/tests/perfstress_tests/create_entity_batch.py
@@ -22,11 +22,11 @@ class CreateEntityBatchTest(_TableTest):
             batch.append(('upsert', self.base_entity))
             batch_size += 1
             if batch_size >= 100:
-                self.table_client.send_batch(batch)
+                self.table_client.submit_transaction(batch)
                 batch = []
                 batch_size = 0
         if batch_size:
-            self.table_client.send_batch(batch)
+            self.table_client.submit_transaction(batch)
 
     async def run_async(self):
         batch_size = 0
@@ -36,11 +36,11 @@ class CreateEntityBatchTest(_TableTest):
             batch.append(('upsert', self.base_entity))
             batch_size += 1
             if batch_size >= 100:
-                await self.async_table_client.send_batch(batch)
+                await self.async_table_client.submit_transaction(batch)
                 batch = []
                 batch_size = 0
         if batch_size:
-            await self.async_table_client.send_batch(batch)
+            await self.async_table_client.submit_transaction(batch)
 
     @staticmethod
     def add_arguments(parser):