+++ sdk/tables/azure-data-tables/tests/perfstress_tests/list_entities.py
@@ -23,11 +23,11 @@ class ListEntitiesTest(_TableTest):
             batch.append(('create', self.base_entity))
             batch_size += 1
             if batch_size >= 100:
-                await self.async_table_client.send_batch(batch)
+                await self.async_table_client.submit_transaction(batch)
                 batch = []
                 batch_size = 0
         if batch_size:
-            await self.async_table_client.send_batch(batch)
+            await self.async_table_client.submit_transaction(batch)
 
     def run_sync(self):
         for _ in self.table_client.list_entities():