+++ sdk/tables/azure-data-tables/tests/test_table_batch.py
@@ -28,7 +28,7 @@ from azure.data.tables import (
     TableEntity,
     EntityProperty,
     UpdateMode,
-    BatchErrorException,
+    TableTransactionError,
     TableServiceClient,
     TableEntity,
     UpdateMode,
@@ -189,7 +189,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
             entity.test5 = datetime.utcnow()
 
             batch = [('create', entity)]
-            transaction_result = self.table.send_batch(batch)
+            transaction_result = self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -231,7 +231,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
             entity.test5 = datetime.utcnow()
 
             batch = [('update', entity, {'mode':UpdateMode.MERGE})]
-            transaction_result = self.table.send_batch(batch)
+            transaction_result = self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -266,7 +266,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
             entity.test2 = u'value1'
 
             batch = [('update', entity)]
-            transaction_result = self.table.send_batch(batch)
+            transaction_result = self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -306,7 +306,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
             entity.test2 = u'value1'
 
             batch = [('update', entity, {'mode': UpdateMode.MERGE})]
-            transaction_result = self.table.send_batch(batch)
+            transaction_result = self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -338,7 +338,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
                 sent_entity,
                 {'etag': etag, 'match_condition':MatchConditions.IfNotModified, 'mode':UpdateMode.REPLACE}
             )]
-            transaction_result = self.table.send_batch(batch)
+            transaction_result = self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -369,8 +369,8 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
                 sent_entity1,
                 {'etag': u'W/"datetime\'2012-06-15T22%3A51%3A44.9662825Z\'"', 'match_condition':MatchConditions.IfNotModified}
             )]
-            with pytest.raises(BatchErrorException):
-                self.table.send_batch(batch)
+            with pytest.raises(TableTransactionError):
+                self.table.submit_transaction(batch)
 
             # Assert
             received_entity = self.table.get_entity(entity['PartitionKey'], entity['RowKey'])
@@ -411,8 +411,8 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
                 {'etag':u'W/"datetime\'2012-06-15T22%3A51%3A44.9662825Z\'"', 'match_condition': MatchConditions.IfNotModified}
             )]
 
-            with pytest.raises(BatchErrorException):
-                self.table.send_batch(batch)
+            with pytest.raises(TableTransactionError):
+                self.table.submit_transaction(batch)
 
             # Assert
             received_entity = self.table.get_entity(entity['PartitionKey'], entity['RowKey'])
@@ -437,7 +437,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
             entity.test5 = datetime.utcnow()
 
             batch = [('upsert', entity, {'mode': UpdateMode.REPLACE})]
-            transaction_result = self.table.send_batch(batch)
+            transaction_result = self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -468,7 +468,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
             entity.test5 = datetime.utcnow()
 
             batch = [('upsert', entity, {'mode': UpdateMode.MERGE})]
-            transaction_result = self.table.send_batch(batch)
+            transaction_result = self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -503,7 +503,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
             assert 3 ==  entity.test3
 
             batch = [('delete', entity)]
-            transaction_result = self.table.send_batch(batch)
+            transaction_result = self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -535,7 +535,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
                 entity.RowKey = str(i)
                 batch.append(('create', entity.copy()))
                 transaction_count += 1
-            transaction_result = self.table.send_batch(batch)
+            transaction_result = self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, transaction_count)
@@ -604,7 +604,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
             batch.append(('upsert', entity.copy(), {'mode': UpdateMode.REPLACE}))
             transaction_count += 1
 
-            transaction_result = self.table.send_batch(batch)
+            transaction_result = self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, transaction_count)
@@ -655,8 +655,8 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
             entity.RowKey = 'batch_all_operations_together-4'
             batch.append(('upsert', entity.copy()))
 
-            resp1 = self.table.send_batch(batch)
-            resp2 = table2.send_batch(batch)
+            resp1 = self.table.submit_transaction(batch)
+            resp2 = table2.submit_transaction(batch)
 
             entities = list(self.table.query_entities("PartitionKey eq '003'"))
             assert 4 ==  len(entities)
@@ -687,8 +687,8 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
 
             # Assert
 
-            with pytest.raises(BatchErrorException):
-                self.table.send_batch(batch)
+            with pytest.raises(TableTransactionError):
+                self.table.submit_transaction(batch)
 
         finally:
             self._tear_down()
@@ -714,7 +714,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
             batch.append(('update', entity.copy()))
 
             with pytest.raises(ValueError):
-                self.table.send_batch(batch)
+                self.table.submit_transaction(batch)
         finally:
             self._tear_down()
 
@@ -728,14 +728,14 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
             self.table.create_entity(entity)
 
             # Act
-            with pytest.raises(BatchErrorException):
+            with pytest.raises(TableTransactionError):
                 batch = []
                 for i in range(0, 101):
                     entity = TableEntity()
                     entity.PartitionKey = 'large'
                     entity.RowKey = 'item{0}'.format(i)
                     batch.append(('create', entity.copy()))
-                self.table.send_batch(batch)
+                self.table.submit_transaction(batch)
 
             # Assert
         finally:
@@ -752,7 +752,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
 
             batch = [('create', entity), ('create', entity2)]
             with pytest.raises(ValueError):
-                self.table.send_batch(batch)
+                self.table.submit_transaction(batch)
 
             # Assert
         finally:
@@ -770,8 +770,8 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
 
             batch = [('create', entity)]
 
-            with pytest.raises(BatchErrorException):
-                resp = tc.send_batch(batch)
+            with pytest.raises(TableTransactionError):
+                resp = tc.submit_transaction(batch)
             # Assert
         finally:
             self._tear_down()
@@ -790,7 +790,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
         batch = [('create', entity)]
 
         with pytest.raises(ClientAuthenticationError):
-            resp = self.table.send_batch(batch)
+            resp = self.table.submit_transaction(batch)
 
     @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @TablesPreparer()
@@ -802,8 +802,8 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
 
             batch = [('delete', entity)]
 
-            with pytest.raises(BatchErrorException):
-                resp = self.table.send_batch(batch)
+            with pytest.raises(TableTransactionError):
+                resp = self.table.submit_transaction(batch)
 
         finally:
             self._tear_down()
@@ -846,7 +846,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
                 entity.RowKey = str(i)
                 batch.append(('create', entity.copy()))
                 transaction_count += 1
-            transaction_result = table.send_batch(batch)
+            transaction_result = table.submit_transaction(batch)
 
             assert transaction_result is not None
 
@@ -878,7 +878,7 @@ class StorageTableBatchTest(AzureTestCase, TableTestCase):
                 batch.append(('create', entity.copy()))
 
             with pytest.raises(RequestTooLargeError):
-                self.table.send_batch(batch)
+                self.table.submit_transaction(batch)
 
         finally:
             self._tear_down()