+++ sdk/tables/azure-data-tables/tests/test_table_batch_cosmos_async.py
@@ -28,7 +28,7 @@ from azure.data.tables import (
     UpdateMode,
     EntityProperty,
     EdmType,
-    BatchErrorException,
+    TableTransactionError,
     RequestTooLargeError
 )
 from azure.data.tables.aio import TableServiceClient
@@ -187,7 +187,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             entity.test5 = datetime.utcnow()
 
             batch = [('create', entity)]
-            transaction_result = await self.table.send_batch(batch)
+            transaction_result = await self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -227,7 +227,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             entity.test5 = datetime.utcnow()
 
             batch = [('update', entity, {'mode':UpdateMode.MERGE})]
-            transaction_result = await self.table.send_batch(batch)
+            transaction_result = await self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -262,7 +262,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             entity.test2 = u'value1'
 
             batch = [('update', entity)]
-            transaction_result = await self.table.send_batch(batch)
+            transaction_result = await self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -300,7 +300,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             entity.test2 = u'value1'
 
             batch = [('update', entity, {'mode': UpdateMode.MERGE})]
-            transaction_result = await self.table.send_batch(batch)
+            transaction_result = await self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -331,7 +331,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
                 sent_entity,
                 {'etag': etag, 'match_condition':MatchConditions.IfNotModified, 'mode':UpdateMode.REPLACE}
             )]
-            transaction_result = await self.table.send_batch(batch)
+            transaction_result = await self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -362,7 +362,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             )]
 
             with pytest.raises(HttpResponseError):
-                await self.table.send_batch(batch)
+                await self.table.submit_transaction(batch)
 
             # Assert
             received_entity = await self.table.get_entity(entity['PartitionKey'], entity['RowKey'])
@@ -386,7 +386,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             entity.test5 = datetime.utcnow()
 
             batch = [('upsert', entity, {'mode': UpdateMode.REPLACE})]
-            transaction_result = await self.table.send_batch(batch)
+            transaction_result = await self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -416,7 +416,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             entity.test5 = datetime.utcnow()
 
             batch = [('upsert', entity, {'mode':UpdateMode.MERGE})]
-            transaction_result = await self.table.send_batch(batch)
+            transaction_result = await self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -450,7 +450,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             assert 3 ==  entity.test3
 
             batch = [('delete', entity)]
-            transaction_result = await self.table.send_batch(batch)
+            transaction_result = await self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
@@ -481,7 +481,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
                 entity.RowKey = str(i)
                 batch.append(('create', entity.copy()))
                 transaction_count += 1
-            transaction_result = await self.table.send_batch(batch)
+            transaction_result = await self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, transaction_count)
@@ -553,7 +553,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             batch.append(('upsert', entity.copy(), {'mode': UpdateMode.REPLACE}))
             transaction_count += 1
 
-            transaction_result = await self.table.send_batch(batch)
+            transaction_result = await self.table.submit_transaction(batch)
 
             # Assert
             self._assert_valid_batch_transaction(transaction_result, transaction_count)
@@ -599,7 +599,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
 
             # Assert
             with pytest.raises(ValueError):
-                await self.table.send_batch(batch)
+                await self.table.submit_transaction(batch)
         finally:
             await self._tear_down()
 
@@ -614,8 +614,8 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
 
             batch = [('create', entity)]
 
-            with pytest.raises(BatchErrorException):
-                resp = await tc.send_batch(batch)
+            with pytest.raises(TableTransactionError):
+                resp = await tc.submit_transaction(batch)
             # Assert
         finally:
             await self._tear_down()
@@ -639,7 +639,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
         batch = [('create', entity)]
 
         with pytest.raises(ClientAuthenticationError):
-            resp = await self.table.send_batch(batch)
+            resp = await self.table.submit_transaction(batch)
 
     @CosmosPreparer()
     async def test_new_delete_nonexistent_entity(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
@@ -649,8 +649,8 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             entity = self._create_random_entity_dict('001', 'batch_negative_1')
 
             batch = [('delete', entity)]
-            with pytest.raises(BatchErrorException):
-                resp = await self.table.send_batch(batch)
+            with pytest.raises(TableTransactionError):
+                resp = await self.table.submit_transaction(batch)
 
         finally:
             await self._tear_down()
@@ -674,7 +674,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
                 batch.append(('create', entity.copy()))
 
             with pytest.raises(RequestTooLargeError):
-                await self.table.send_batch(batch)
+                await self.table.submit_transaction(batch)
 
         finally:
             await self._tear_down()